Toss Refactoring 완료

+++ back/src/main/java/org/zerock/mallapi/controller/CartController.java
@@ -3,14 +3,18 @@ package org.zerock.mallapi.controller;
 import java.security.Principal;
 import java.util.List;
 
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import org.zerock.mallapi.dto.*;
 
+import org.zerock.mallapi.exception.ErrorCode;
 import org.zerock.mallapi.service.CartService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.zerock.mallapi.util.GeneralException;
 
 @RestController
 @RequiredArgsConstructor
@@ -20,13 +24,23 @@ public class CartController {
 
     private final CartService cartService;
 
-    @PreAuthorize("#itemDTO.email == authentication.name")
+//    @PreAuthorize("#itemDTO.email == authentication.name")
     @PostMapping("/change")
-    public DataResponseDTO<List<CartItemListDTO>> changeCart(@RequestBody CartItemDTO itemDTO){
+    public DataResponseDTO<List<CartItemListDTO>> changeCart(@RequestBody CartItemDTO itemDTO, Authentication authentication) {
+
+        if (!authentication.getName().equals(itemDTO.getEmail())) {
+            throw new GeneralException(ErrorCode.INVALID_TOKEN, "이메일 불일치: 인증된 사용자 아님");
+//            throw new AccessDeniedException("이메일 불일치: 인증된 사용자 아님");
+        }
+
+        itemDTO.setEmail(authentication.getName());
 
         log.info("itemDTO............................." + itemDTO);
 
-        if(itemDTO.getQty() <= 0) {
+        log.info("itemDTO.email = " + itemDTO.getEmail());
+        log.info("auth.name = " + authentication.getName());
+
+        if (itemDTO.getQty() <= 0) {
             return DataResponseDTO.of(cartService.remove(itemDTO.getCino()));
         }
